LayoutTests/svg/animations/animate-text-nested-transforms-expected.txt

SVG 1.1 dynamic animation tests

This is some text.
test to determine whether auto-rotate animateMotion path animations pre-muliply or post-multiply animation transform matrix

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rootSVGElement.getBBox().x is 118.65
PASS rootSVGElement.getBBox().y is -161.03
PASS rootSVGElement.getBBox().x is 366.89
PASS rootSVGElement.getBBox().y is 549.93
PASS successfullyParsed is true

TEST COMPLETE
